+++ web/src/components/App.js
@@ -7,13 +7,13 @@ import DevProvider from './DevProvider';
 import NoMatch from './NoMatch';
 import ActivitiesStart from '../containers/ActivitiesStart';
 import ActivitiesList from '../containers/ActivitiesList';
+import ActivityApproach from '../containers/ActivityApproach';
+import ActivityGoals from '../containers/ActivityGoals';
 import ActivityOverview from '../containers/ActivityOverview';
 import ApdOverview from '../containers/ApdOverview';
 import StateContacts from '../containers/StateContacts';
 import StateStart from '../containers/StateStart';
 import StatePersonnel from '../containers/StatePersonnel';
-import Approach from '../containers/Approach';
-import GoalsAndObjectives from '../containers/GoalsAndObjectives';
 
 const Wrapper = process.env.NODE_ENV !== 'production' ? DevProvider : Provider;
 
@@ -31,15 +31,15 @@ const App = () => (
     <Container is="main">
       <Switch>
         <Route exact path="/" component={Demo} />
-        <Route path="/activities-start" component={ActivitiesStart} />
         <Route path="/activities-list" component={ActivitiesList} />
+        <Route path="/activities-start" component={ActivitiesStart} />
+        <Route path="/activity-approach" component={ActivityApproach} />
+        <Route path="/activity-goals" component={ActivityGoals} />
         <Route path="/activity-overview" component={ActivityOverview} />
         <Route path="/apd-overview" component={ApdOverview} />
-        <Route path="/state-start" component={StateStart} />
         <Route path="/state-contacts" component={StateContacts} />
         <Route path="/state-personnel" component={StatePersonnel} />
-        <Route path="/approach" component={Approach} />
-        <Route path="/goals-and-objectives" component={GoalsAndObjectives} />
+        <Route path="/state-start" component={StateStart} />
         <Route component={NoMatch} />
       </Switch>
     </Container>
