+++ web/src/containers/ActivityGoals.js
@@ -6,10 +6,10 @@ import { Box, Heading } from 'rebass';
 import { bindActionCreators } from 'redux';
 
 import FormLogger from '../util/formLogger';
-import FormStateContacts from '../components/FormStateContacts';
+import FormActivityGoals from '../components/FormActivityGoals';
 import PageNavButtons from '../components/PageNavButtons';
 
-class StateContacts extends Component {
+class ActivityGoals extends Component {
   showResults = data => {
     console.log(data);
   };
@@ -20,19 +20,25 @@ class StateContacts extends Component {
     return (
       <Box py={4}>
         <FormLogger />
-        <Heading mb={3}>Review your state contact information</Heading>
-        <FormStateContacts onSubmit={this.showResults} />
-        <PageNavButtons goTo={goTo} prev="/state-start" next="/apd-overview" />
+        <Heading mb={3}>
+          Add goals and objectives for <em>Administration</em>
+        </Heading>
+        <FormActivityGoals onSubmit={this.showResults} />
+        <PageNavButtons
+          goTo={goTo}
+          prev="/activity-overview"
+          next="/activity-approach"
+        />
       </Box>
     );
   }
 }
 
-StateContacts.propTypes = {
+ActivityGoals.propTypes = {
   goTo: PropTypes.func.isRequired
 };
 
 const mapDispatchToProps = dispatch =>
   bindActionCreators({ goTo: path => push(path) }, dispatch);
 
-export default connect(null, mapDispatchToProps)(StateContacts);
+export default connect(null, mapDispatchToProps)(ActivityGoals);
